feat: Handle error and redirect to home page when clicking "Go Home" button
The code changes in `error.vue` add a button with the text "Go Home" and a click event handler called `handleError`. When the button is clicked, the `handleError` function is called, which clears the error and redirects the user to the home page. This improves the user experience by providing a way to easily navigate back to the home page when encountering an error.